Fixed MsJobSubmitter -- checks for correct access to a project.

+++ MsUploadQueueWS/trunk/src/org/yeastrc/jqs/queue/ws/MsJobResourceClient.java
@@ -89,7 +89,7 @@ public class MsJobResourceClient {
 	
 	/*
 	 * Equivalent curl commands -- 
-	 * JSON INPUT: curl -u <username> -i -X POST -H "Accept:text/plain" -H 'Content-Type: application/json' -d '{"userName":"vsharma", "projectId":"24", "dataDirectory":"/test/data", "pipeline":"MACOSS", "date":"2010-03-29"}' "http://localhost:8080/msdapl_queue/services/msjob/add"
+	 * JSON INPUT: curl -u <username> -i -X POST -H "Accept:text/plain" -H 'Content-Type: application/json' -d '{"projectId":"24", "dataDirectory":"/test/data", "pipeline":"MACOSS", "date":"2010-03-29"}' "http://localhost:8080/msdapl_queue/services/msjob/add"
 	 */
 	public int addJob(MsJob job, String username, String password) {
 		
@@ -116,7 +116,7 @@ public class MsJobResourceClient {
 	
 	/*
 	 * Equivalent curl commands -- 
-	 * curl -u <username> -i -X POST -H "Accept:text/plain" "http://localhost:8080/msdapl_queue/services/msjob/add?user=vsharma&projectId=24&dir=/data/test&pipeline=MACOSS&date=09/24/10&instrument=LTQ&Id=9606&comments=some%20comments"
+	 * curl -u <username> -i -X POST -H "Accept:text/plain" "http://localhost:8080/msdapl_queue/services/msjob/add?projectId=24&dir=/data/test&pipeline=MACOSS&date=09/24/10&instrument=LTQ&Id=9606&comments=some%20comments"
 	 */
 	public int addJobQueryParam(MsJob job, String username, String password) {
 
@@ -183,22 +183,24 @@ public class MsJobResourceClient {
 		
 		MsJobResourceClient client = MsJobResourceClient.getInstance();
 		
-		String username = "vsharma";
-		String password = "Bits2@";
+		String username = "test";
+		String password = "user";
 		
 		MsJob job = new MsJob();
-		job.setSubmitterName("vsharma");
+//		job.setSubmitterName("vsharma");
 		job.setProjectId(24);
 		job.setDataDirectory("/test/dir");
 		job.setPipeline("TPP");
 		job.setDate(new Date());
 		int jobId = client.addJob(job, username, password);
 		
-		client.getJob(jobId);
-		
-		client.getJobStatus(jobId);
-		
-		client.delete(jobId, username, password);
+		if(jobId > 0) {
+			client.getJob(jobId);
+
+			client.getJobStatus(jobId);
+
+			client.delete(jobId, username, password);
+		}
 	}
 
 }
